feat: replace update contact dialog with reusable ContactForm and update ticket detail view integration
