+++ apps/launcher/src/shared/assets.ts
@@ -3,21 +3,37 @@ export const INTERAONE_LOGO_SVG = `
   <path d="M12 2.8l1.78 5.42L19.2 10l-5.42 1.78L12 17.2l-1.78-5.42L4.8 10l5.42-1.78L12 2.8Z"></path>
 </svg>`;
 
-export const INTERAONE_FLOATING_LOGO_SVG = `
-<svg width="34" height="34" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
-  <g clip-path="url(#clip0)">
-    <path
-      fill-rule="evenodd"
-      clip-rule="evenodd"
-      d="M5.52569 -0.158142L4.35475 3.35469L0.841919 4.52563V5.47431L4.35475 6.64526L5.52569 10.1581H6.47437L7.64532 6.64526L11.1581 5.47431V4.52563L9.58126 4L12.9999 4V6.5L8.99994 8L7.49994 12H4.49994L3.02379 8.06377C2.99998 8.54193 2.99999 9.10377 3 9.75869V13.2407C2.99999 14.0457 2.99998 14.7101 3.04419 15.2513C3.09012 15.8134 3.18868 16.3301 3.43597 16.8154C3.81947 17.568 4.43139 18.18 5.18404 18.5635C5.66937 18.8108 6.18608 18.9093 6.74817 18.9552C7.28938 18.9995 7.95374 18.9995 8.75872 18.9994H8.96859L12 22.5368L15.0314 18.9994H15.2413C16.0463 18.9995 16.7106 18.9995 17.2518 18.9552C17.8139 18.9093 18.3306 18.8108 18.816 18.5635C19.5686 18.18 20.1805 17.5681 20.564 16.8154C20.8113 16.3301 20.9099 15.8134 20.9558 15.2513C21 14.7101 21 14.0457 21 13.2408V9.75868C21 8.95372 21 8.28936 20.9558 7.74818C20.9099 7.18608 20.8113 6.66937 20.564 6.18404C20.1805 5.43139 19.5686 4.81947 18.816 4.43597C18.3306 4.18868 17.8139 4.09012 17.2518 4.04419C16.7106 3.99998 16.0463 3.99999 15.2413 4L9.58126 4L7.64532 3.35469L6.47437 -0.158142H5.52569ZM13.5257 8.84186L12.8547 10.8547L10.8419 11.5256V12.4743L12.8547 13.1453L13.5257 15.1581H14.4744L15.1453 13.1453L17.1581 12.4743V11.5256L15.1453 10.8547L14.4744 8.84186H13.5257Z"
-      fill="white"
-    />
-  </g>
-  <defs>
-    <clipPath id="clip0">
-      <rect width="24" height="24"/>
-    </clipPath>
-  </defs>
-</svg>
+// export const INTERAONE_FLOATING_LOGO_SVG = `
+// <svg width="34" height="34" viewBox="0 0 24 24" fill="none" xmlns="http://www.w3.org/2000/svg">
+//   <g clip-path="url(#clip0)">
+//     <path
+//       fill-rule="evenodd"
+//       clip-rule="evenodd"
+//       d="M5.52569 -0.158142L4.35475 3.35469L0.841919 4.52563V5.47431L4.35475 6.64526L5.52569 10.1581H6.47437L7.64532 6.64526L11.1581 5.47431V4.52563L9.58126 4L12.9999 4V6.5L8.99994 8L7.49994 12H4.49994L3.02379 8.06377C2.99998 8.54193 2.99999 9.10377 3 9.75869V13.2407C2.99999 14.0457 2.99998 14.7101 3.04419 15.2513C3.09012 15.8134 3.18868 16.3301 3.43597 16.8154C3.81947 17.568 4.43139 18.18 5.18404 18.5635C5.66937 18.8108 6.18608 18.9093 6.74817 18.9552C7.28938 18.9995 7.95374 18.9995 8.75872 18.9994H8.96859L12 22.5368L15.0314 18.9994H15.2413C16.0463 18.9995 16.7106 18.9995 17.2518 18.9552C17.8139 18.9093 18.3306 18.8108 18.816 18.5635C19.5686 18.18 20.1805 17.5681 20.564 16.8154C20.8113 16.3301 20.9099 15.8134 20.9558 15.2513C21 14.7101 21 14.0457 21 13.2408V9.75868C21 8.95372 21 8.28936 20.9558 7.74818C20.9099 7.18608 20.8113 6.66937 20.564 6.18404C20.1805 5.43139 19.5686 4.81947 18.816 4.43597C18.3306 4.18868 17.8139 4.09012 17.2518 4.04419C16.7106 3.99998 16.0463 3.99999 15.2413 4L9.58126 4L7.64532 3.35469L6.47437 -0.158142H5.52569ZM13.5257 8.84186L12.8547 10.8547L10.8419 11.5256V12.4743L12.8547 13.1453L13.5257 15.1581H14.4744L15.1453 13.1453L17.1581 12.4743V11.5256L15.1453 10.8547L14.4744 8.84186H13.5257Z"
+//       fill="white"
+//     />
+//   </g>
+//   <defs>
+//     <clipPath id="clip0">
+//       <rect width="24" height="24"/>
+//     </clipPath>
+//   </defs>
+// </svg>
 
+// `;
+
+export const INTERAONE_FLOATING_LOGO_SVG = `
+<svg
+  width="28"
+  height="28"
+  viewBox="0 0 24 24"
+  fill="currentColor"
+  xmlns="http://www.w3.org/2000/svg"
+>
+  <path
+    fill-rule="evenodd"
+    clip-rule="evenodd"
+    d="M8.7587 4H15.2413C16.0463 3.99999 16.7106 3.99998 17.2518 4.04419C17.8139 4.09012 18.3306 4.18868 18.816 4.43598C19.5686 4.81947 20.1805 5.43139 20.564 6.18404C20.8113 6.66937 20.9099 7.18608 20.9558 7.74818C21 8.28937 21 8.95372 21 9.75868V13.2408C21 14.0457 21 14.7101 20.9558 15.2513C20.9099 15.8134 20.8113 16.3301 20.564 16.8154C20.1805 17.568 19.5686 18.18 18.816 18.5635C18.3306 18.8108 17.8139 18.9093 17.2518 18.9552C16.7106 18.9995 16.0463 18.9995 15.2413 18.9994H15.0314L12 22.5368L8.96859 18.9994H8.75872C7.95374 18.9995 7.28938 18.9995 6.74818 18.9552C6.18608 18.9093 5.66937 18.8108 5.18404 18.5635C4.43139 18.18 3.81947 17.568 3.43598 16.8154C3.18868 16.3301 3.09012 15.8134 3.04419 15.2513C2.99998 14.7101 2.99999 14.0457 3 13.2408V9.7587C2.99999 8.95374 2.99998 8.28937 3.04419 7.74818C3.09012 7.18608 3.18868 6.66937 3.43598 6.18404C3.81947 5.43139 4.43139 4.81947 5.18404 4.43598C5.66937 4.18868 6.18608 4.09012 6.74818 4.04419C7.28937 3.99998 7.95374 3.99999 8.7587 4ZM9.00003 7.58579L12.4142 11L9.00003 14.4142L7.58582 13L9.58582 11L7.58582 9L9.00003 7.58579ZM16 13H12V15H16V13Z"
+  />
+</svg>
 `;
